refactor: Изменение структуры проекта (перенос ML-component внутрь основного модуля)
