+++ src/routers/generate_image_button.py
