+++ src/routers/magic_button.py
