+++ src/routers/remove_noise_button.py
@@ -1,14 +1,13 @@
-from aiogram import Router, html, F
+from aiogram import Router, F
 from aiogram.fsm.context import FSMContext
 from aiogram.types import Message, CallbackQuery
 import supabase as sb
 import io
 
-from ..keyboards_buttons import menu_buttons, ButtonText
+from ..keyboards_buttons import menu_buttons
 from .button_states import Form, DelNoise_States
 from tempfile import mkdtemp
-from ML_component import main_model
-
+from ..ML_component import main_model
 
 remove_noise = Router()
